Fix reproduced main CI failures in termination events and codex wrapper test (#77)
* Fix CI failures in PG termination events and codex wrapper test

* Update generated module inventory for CI

---------

+++ src/server/routes/termination_events.rs
@@ -229,7 +229,7 @@ async fn list_termination_events_pg(
                 "reason_text": row.try_get::<Option<String>, _>("reason_text").map_err(|error| format!("{error}"))?,
                 "probe_snapshot": row.try_get::<Option<String>, _>("probe_snapshot").map_err(|error| format!("{error}"))?,
                 "last_offset": row.try_get::<Option<i64>, _>("last_offset").map_err(|error| format!("{error}"))?,
-                "tmux_alive": row.try_get::<Option<i32>, _>("tmux_alive").map_err(|error| format!("{error}"))?.map(|value| value != 0),
+                "tmux_alive": row.try_get::<Option<i64>, _>("tmux_alive").map_err(|error| format!("{error}"))?.map(|value| value != 0),
                 "created_at": row.try_get::<String, _>("created_at").map_err(|error| format!("{error}"))?,
             }))
         })
